Fixed some problems and design issues
Dialog dismiss function was added to eliminate memory leakage that
caused app crash.
Design of the sign in layout.
Removed the coordinates and date/time from map layout as it is not
needed for the user.
Edited the MainActivity class to remove the map layout components and
some other things that were not needed.